Added (Add|Delete|Populate) functionality
Works very similar to our User code, which is a pain as the tests are
quite tightly coupled with the (Add|Delete|Populate|Exists) functions.
This meant that this commit is an absolute beast and as such, I've had
to go through it a number of times to ensure that all the code should
work as expected.